# docs/seller-swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "E-Commerce API",
    "version": "1.0.0",
    "title": "E-Commerce API",
    "basePath": "/v1",

    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "basePath": "/v1",
  "schemes": ["http"],
  "produces": ["application/json"],
  "consumes": ["application/json"],
  "paths": {
...
    "/sellers/products": {
      "post": {
        "tags": ["Seller Products"],
        "summary": "Create a new product",
        "description": "Create a new product",
        "operationId": "createProduct",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be created",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "description": {
                  "type": "string"
                },
                "category": {
                  "type": "string"
                },
                "defaultImage": {
                  "type": "string"
                }
              }
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },

      "get": {
        "tags": ["Seller Products"],
        "summary": "Get all products",
        "description": "Get all products",
        "operationId": "getSellerProducts",
        "produces": ["application/json"],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/sellers/products/{productId}/variants": {
      "post": {
        "tags": ["Seller Products"],
        "description": "Create a new product variant",
        "summary": "Create a new product variant",
        "operationId": "createProductVariant",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be created",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "stock": {
                  "type": "integer"
                },
                "imageUrl": {
                  "type": "string"
                },
                "variationOptions": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "name": {
                        "type": "string"
                      },
                      "value": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
    "/sellers/products/{productId}/variants/{variantId}": {
      "get": {
        "tags": ["Seller Products"],
        "description": "Get product variant by id",
        "summary": "Get product variant by id",
        "operationId": "getProductVariantById",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer"
          },
          {
            "name": "variantId",
            "in": "path",
            "description": "ID of product variant to return",
            "required": true,
            "type": "integer"
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },

      "delete": {
        "tags": ["Seller Products"],
        "description": "Delete product variant by id",
        "summary": "Delete product variant by id",
        "operationId": "deleteProductVariantById",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer"
          },
          {
            "name": "variantId",
            "in": "path",
            "description": "ID of product variant to return",
            "required": true,
            "type": "integer"
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      },

      "put": {
        "tags": ["Seller Products"],
        "description": "Update product variant by id",
        "summary": "Update product variant by id",
        "operationId": "updateProductVariantById",
        "consumes": ["application/json"],
        "produces": ["application/json"],

        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "description": "ID of product to return",
            "required": true,
            "type": "integer"
          },
          {
            "name": "variantId",
            "in": "path",
            "description": "ID of product variant to return",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Product object that needs to be updated",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "price": {
                  "type": "number"
                },
                "stock": {
                  "type": "integer"
                },
                "imageUrl": {
                  "type": "string"
                }
              }
            }
          }
        ],

        "responses": {
          "200": {
            "description": "Successful operation"
          }
        }
      }
    },
...
  },
...
}